Moved helper functions to const.py
